- Split between data generation and service

+++ src/main/java/WoWSFT/model/gameparams/consumable/Consumable.java
@@ -17,8 +17,8 @@ public class Consumable
 
     private boolean canBuy;
     private LinkedHashMap<String, Boolean> canBuyCustom;
-    private float costCR;
-    private float costGold;
+    private double costCR;
+    private double costGold;
     private boolean freeOfCharge;
     private long id;
     private String index;
