+++ src/main/java/WoWSFT/model/gameparams/ship/component/engine/Engine.java
@@ -9,16 +9,16 @@ import lombok.Data;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class Engine
 {
-    private float backwardEngineForsag;
-    private float backwardEngineForsagMaxSpeed;
-    private float backwardEngineUpTime;
-    private float backwardSpeedOnFlood;
-    private float damagedEnginePowerMultiplier;
-    private float damagedEnginePowerTimeMultiplier;
-    private float forwardEngineForsag;
-    private float forwardEngineForsagMaxSpeed;
-    private float forwardEngineUpTime;
-    private float forwardSpeedOnFlood;
-    private float histEnginePower;
-    private float speedCoef;
+    private double backwardEngineForsag;
+    private double backwardEngineForsagMaxSpeed;
+    private double backwardEngineUpTime;
+    private double backwardSpeedOnFlood;
+    private double damagedEnginePowerMultiplier;
+    private double damagedEnginePowerTimeMultiplier;
+    private double forwardEngineForsag;
+    private double forwardEngineForsagMaxSpeed;
+    private double forwardEngineUpTime;
+    private double forwardSpeedOnFlood;
+    private double histEnginePower;
+    private double speedCoef;
 }
