+++ src/main/java/WoWSFT/model/gameparams/ship/upgrades/ShipUpgradeInfo.java
@@ -29,7 +29,7 @@ public class ShipUpgradeInfo
     private int costSaleGold;
     private int costXP;
     private List<Object> lockedConfig;
-    private float prepareTimeFactor;
+    private double prepareTimeFactor;
     private int value;
 
     {
